requirements.txt
langchain==0.0.157
python-dotenv==1.0.0
Requests==2.29.0
weaviate_client==3.16.2
websockets==11.0.2
